+++ lib/amazing_print/formatters/dir_formatter.rb
@@ -11,6 +11,7 @@ module AmazingPrint
       attr_reader :dir, :inspector, :options
 
       def initialize(dir, inspector)
+        super()
         @dir = dir
         @inspector = inspector
         @options = inspector.options