+++ spec/active_record_helper.rb
@@ -27,8 +27,11 @@ if ExtVerifier.has_rails?
 
   # Create models
   class User < ActiveRecord::Base; has_many :emails; end
+
   class SubUser < User; end
+
   class Email < ActiveRecord::Base; belongs_to :user; end
+
   class TableFreeModel
     include ::ActiveModel::Validations
     attr_reader(:name)