+++ spec/ext/mongo_mapper_spec.rb
@@ -1,5 +1,7 @@
 # frozen_string_literal: true
 
+# rubocop:disable Lint/ConstantDefinitionInBlock
+
 require 'spec_helper'
 
 RSpec.describe 'AmazingPrint/MongoMapper', skip: -> { !ExtVerifier.has_mongo_mapper? }.call do
@@ -27,11 +29,11 @@ RSpec.describe 'AmazingPrint/MongoMapper', skip: -> { !ExtVerifier.has_mongo_map
     # before { @ap.options[:raw] = true }
     before { @ap = AmazingPrint::Inspector.new(plain: true, sort_keys: true, raw: true) }
 
-    it 'should print class instance' do
+    it 'prints class instance' do
       user = MongoUser.new(first_name: 'Al', last_name: 'Capone')
 
       out = @ap.send(:awesome, user)
-               .gsub(/#\<Proc:.+?\>/, 'amazing_print_PROC_STUB')
+               .gsub(/#<Proc:.+?>/, 'amazing_print_PROC_STUB')
                .gsub(/BSON::ObjectId\('[\da-f]+?'\)/, "BSON::ObjectId('123456789')")
 
       str = if MongoMapper::Version >= '0.13'
@@ -129,7 +131,7 @@ RSpec.describe 'AmazingPrint/MongoMapper', skip: -> { !ExtVerifier.has_mongo_map
       expect(out).to be_similar_to(str)
     end
 
-    it 'should print the class' do
+    it 'prints the class' do
       expect(@ap.send(:awesome, MongoUser)).to eq <<~EOS.strip
         class MongoUser < Object {
                    "_id" => :object_id,
@@ -139,7 +141,7 @@ RSpec.describe 'AmazingPrint/MongoMapper', skip: -> { !ExtVerifier.has_mongo_map
       EOS
     end
 
-    it 'should print the class when type is undefined' do
+    it 'prints the class when type is undefined' do
       class Chamelion
         include MongoMapper::Document
         key :last_attribute
@@ -178,7 +180,7 @@ RSpec.describe 'AmazingPrint/MongoMapper', skip: -> { !ExtVerifier.has_mongo_map
     end
 
     describe 'with show associations turned off (default)' do
-      it 'should render the class as normal' do
+      it 'renders the class as normal' do
         expect(@ap.send(:awesome, Parent)).to eq <<~EOS.strip
           class Parent < Object {
                "_id" => :object_id,
@@ -187,7 +189,7 @@ RSpec.describe 'AmazingPrint/MongoMapper', skip: -> { !ExtVerifier.has_mongo_map
         EOS
       end
 
-      it 'should render an instance as normal' do
+      it 'renders an instance as normal' do
         parent = Parent.new(name: 'test')
         out = @ap.send(:awesome, parent)
         str = <<~EOS.strip
@@ -201,11 +203,11 @@ RSpec.describe 'AmazingPrint/MongoMapper', skip: -> { !ExtVerifier.has_mongo_map
     end
 
     describe 'with show associations turned on and inline embedded turned off' do
-      before :each do
+      before do
         @ap = AmazingPrint::Inspector.new(plain: true, mongo_mapper: { show_associations: true })
       end
 
-      it 'should render the class with associations shown' do
+      it 'renders the class with associations shown' do
         expect(@ap.send(:awesome, Parent)).to eq <<~EOS.strip
           class Parent < Object {
                   "_id" => :object_id,
@@ -216,7 +218,7 @@ RSpec.describe 'AmazingPrint/MongoMapper', skip: -> { !ExtVerifier.has_mongo_map
         EOS
       end
 
-      it 'should render an instance with associations shown' do
+      it 'renders an instance with associations shown' do
         parent = Parent.new(name: 'test')
         out = @ap.send(:awesome, parent)
         str = <<~EOS.strip
@@ -232,7 +234,7 @@ RSpec.describe 'AmazingPrint/MongoMapper', skip: -> { !ExtVerifier.has_mongo_map
     end
 
     describe 'with show associations turned on and inline embedded turned on' do
-      before :each do
+      before do
         @ap = AmazingPrint::Inspector.new plain: true,
                                           mongo_mapper: {
                                             show_associations: true,
@@ -240,7 +242,7 @@ RSpec.describe 'AmazingPrint/MongoMapper', skip: -> { !ExtVerifier.has_mongo_map
                                           }
       end
 
-      it 'should render an instance with associations shown and embeds there' do
+      it 'renders an instance with associations shown and embeds there' do
         parent = Parent.new(name: 'test', child: Child.new(data: 5))
         out = @ap.send(:awesome, parent)
         str = <<~EOS.strip
@@ -259,3 +261,5 @@ RSpec.describe 'AmazingPrint/MongoMapper', skip: -> { !ExtVerifier.has_mongo_map
     end
   end
 end
+
+# rubocop:enable Lint/ConstantDefinitionInBlock