+++ CHANGELOG.md
@@ -1,6 +1,28 @@
-## unreleased
+## Unreleased
   - Drop support for Ruby 2.3 and 2.4 as well as JRuby 9.1
   - Add File/Dir formatters for mswin platform
+
+## v1.4.0
+  - Support loading config from `$XDG_CONFIG_HOME/aprc` - #63
+  - Remove support for Rails 5.1 #75
+  - Update AR specs for Ruby 2.6.7 #76
+  - Load .aprc configs only once. #74
+  - Add XDG config support #77
+  - Rubocop updates #79
+  - Update Irb integration for v1.2.6+ #81
+
+## v1.3.0
+  - Fix HTML escaping problems #53
+  - Update test suite for Ruby 2.7.2 and JRuby #61
+  - Add ActionView spec for html_safe #65
+  - Add support for Rails 6.1 #68
+  - Update specs for Ruby 3.0 #69
+
+## v1.2.2
+  - Support Ruby 3.0 / IRB 1.2.6 - #57
+  - Fix FrozenError - #51
+  - Drop support for Ruby 2.3 and 2.4 as well as JRuby 9.1 - #46
+  - Add passing of `options` to `Logger#ap` - #55
 
 ## v1.2.1
   - Correctly print active_model_errors for models that don't have tables - #42 by sahglie